requirements.txt
numpy
pyyaml
tensorboardX
munch
scipy
matplotlib
trimesh
huepy
pillow
pycolmap
plyfile
tqdm
scikit-learn
opencv-python-headless
torch
torchvision
lpips
scikit-image
click
